refactor: Renamed logging module as log module
- prevents clash with "logging" phrase in info log for tests.
- updated references to logging module to log.
